+++ src/App.tsx
@@ -5,13 +5,13 @@ import MainPage from "./components/pages/MainPage";
 import { Routes, Route } from "react-router-dom";
 import AttachmentPage from "./components/pages/AttachmentPage";
 import DashboardPage from "./components/pages/DashboardPage";
-import Footer from "./components/sections/Footer";
+import Footer from "./components/sections/footer/Footer";
 import DetailPage from "./components/pages/DetailPage";
 import { useSelector } from "react-redux";
-import FloatingMenu from "./components/organisms/FloatingMenu";
+import FloatingMenu from "./components/molecules/FloatingMenu";
 import Wrapper from "./components/wrapper/Wrapper";
 import GlobalStyle from "./style/GlobalStyles";
-import Navigation from "./components/sections/Navigation";
+import Navigation from "./components/sections/navigation/Navigation";
 
 function App() {
   const isDarkMode = useSelector((state: { isDarkModeSlice: boolean }) => state.isDarkModeSlice);
