+++ src/components/organisms/attachment/OsList.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import Span from "../atoms/Span";
+import Span from "../../atoms/Span";
 import styled from "styled-components";
 import { useDispatch, useSelector } from "react-redux";
-import { setSelectedOsIndex } from "../../store/reducer/selectedOsIndexSlice";
+import { setSelectedOsIndex } from "../../../store/reducer/selectedOsIndexSlice";
 import { AiFillWindows, AiFillApple, AiFillAndroid } from "react-icons/ai";
 import { SiIos } from "react-icons/si";
-import Icon from "../atoms/Icon";
-import { lightTheme } from "../../style/Theme";
+import Icon from "../../atoms/Icon";
+import { lightTheme } from "../../../style/Theme";
 
 const OsListBox = styled.ul`
   margin: 0 auto;
