+++ src/components/organisms/attachment/ThreeImages.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import ImageCard from "../molecules/ImageCard";
+import ImageCard from "../../molecules/attachment/ImageCard";
 
 const ThreeImagesBox = styled.div`
   padding: 0 10px;
