+++ src/components/organisms/dashboard/DatePickerCalendar.tsx
@@ -3,8 +3,8 @@ import DatePicker from "react-datepicker";
 import styled from "styled-components";
 import { useDispatch, useSelector } from "react-redux";
 import "react-datepicker/dist/react-datepicker.css"; // react-datepicker의 CSS 파일을 import
-import { setLimitTime } from "../../store/reducer/limitTimeSlice";
-import { AnalyzedMessage } from "../../@types/index.d";
+import { setLimitTime } from "../../../store/reducer/limitTimeSlice";
+import { AnalyzedMessage } from "../../../@types/index.d";
 
 const DatePickerInputContainer = styled.div`
   color: ${(props) => props.theme.mainBlue};
