+++ src/components/pages/AttachmentPage.tsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from "react";
 import styled from "styled-components";
 
-import AttachmentSection from "../sections/AttachmentSection";
-import AttachmentDescriptionSection from "../sections/AttachDescriptionSection";
+import AttachmentSection from "../sections/attachment/AttachmentSection";
+import AttachmentDescriptionSection from "../sections/attachment/AttachDescriptionSection";
 import scrollToEvent from "../../module/common/scrollToEvent";
 import { FlexColumnCenterDiv } from "../styleComponents/FlexDiv";
 import { useDispatch } from "react-redux";
-import { pushNewlyAttachedFiles } from "../../store/reducer/attachedFileListSlice";
+import { pushNewlyAttachedFiles } from "../../store/reducer/attachment/attachedFileListSlice";
 import { VscNewFile } from "react-icons/vsc";
 import Span from "../atoms/Span";
