+++ src/components/pages/DetailPage.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 import scrollToEvent from "../../module/common/scrollToEvent";
-import DashboardSideMenu from "../sections/DashboardSideMenu";
-import ModalGraph from "../organisms/ModalGraph";
+import DashboardSideMenu from "../sections/dashboard/DashboardSideMenu";
+import ModalGraph from "../organisms/dashboard/ModalGraph";
 import ChatRoomCompareGraph from "../molecules/graphs/ChatRoomCompareGraph";
 import ChatVolumeByPeriodGraph from "../molecules/graphs/ChatVolumeByPeriodGraph";
 import ChatRateGraph from "../molecules/graphs/ChatRateGraph";
