+++ src/components/sections/dashboard/DashboardSideMenu.tsx
@@ -3,17 +3,17 @@ import styled from "styled-components";
 import SummaryPieGraph, {
   getTotalChatCounts,
   getTwoLettersFromSpeakers,
-} from "../molecules/graphs/SummaryPieGraph";
+} from "../../molecules/graphs/SummaryPieGraph";
 import { useDispatch, useSelector } from "react-redux";
-import { AnalyzedMessage, ChatTimes } from "../../@types/index.d";
-import { getChatTimes, getSpeakers } from "../../module/common/getProperties";
-import Span from "../atoms/Span";
-import Paragraph from "../atoms/Paragraph";
-import { setSelectedChatRoomIndex } from "../../store/reducer/selectedRoomIndexSlice";
+import { AnalyzedMessage, ChatTimes } from "../../../@types/index.d";
+import { getChatTimes, getSpeakers } from "../../../module/common/getProperties";
+import Span from "../../atoms/Span";
+import Paragraph from "../../atoms/Paragraph";
+import { setSelectedChatRoomIndex } from "../../../store/reducer/selectedRoomIndexSlice";
 import { Link } from "react-router-dom";
-import { setSelectedSpeakerIndex } from "../../store/reducer/selectedSpeakerIndexSlice";
+import { setSelectedSpeakerIndex } from "../../../store/reducer/selectedSpeakerIndexSlice";
 
-import { setIsSideMenuChatRoom } from "../../store/reducer/isSideMenuChatRoomSelectSlice";
+import { setIsSideMenuChatRoom } from "../../../store/reducer/isSideMenuChatRoomSelectSlice";
 
 const DashboardLayoutBox = styled.div`
   position: sticky;
