+++ src/components/sections/main/MainDescriptionSection.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import styled from "styled-components";
-import H2 from "../atoms/H2";
-import H3 from "../atoms/H3";
-import Img from "../atoms/Img";
-import Paragraph from "../atoms/Paragraph";
-import { FlexCenterDiv, FlexColumnCenterDiv } from "../styleComponents/FlexDiv";
+import H2 from "../../atoms/H2";
+import H3 from "../../atoms/H3";
+import Img from "../../atoms/Img";
+import Paragraph from "../../atoms/Paragraph";
+import { FlexCenterDiv, FlexColumnCenterDiv } from "../../styleComponents/FlexDiv";
 
 const Container = styled.div`
   margin: 100px 0px 0px 0px;
